ACTUAL 1.7.10 compatibility here
Had to replace a lot of getAABBPool and
world.getWorldVec3Pool().getVecFromPool stuff. Otherwise, a lot of
things, like wayfarer staves and guns would CRASH Minecraft! I also
removed the modstats stuff (since there was a change to the md5 hashing
function) and added in libsandstone and JToml code directly (because
they refused to soft depend during compile and runtime). It's fugly,
it's messy... but it works flawlessly in 1.7.10 and that's all I care
about right now.